requirements.txt
asgiref==3.7.2
black==23.7.0
click==8.1.5
Django==4.2.3
django-cors-headers==4.2.0
django-uuslug==2.0.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.7
inflection==0.5.1
mypy-extensions==1.0.0
mysql-client==0.0.1
mysqlclient==2.2.0
packaging==23.1
pathspec==0.11.1
Pillow==10.0.0
platformdirs==3.9.1
PyJWT==2.8.0
python-dotenv==1.0.0
python-slugify==8.0.1
pytz==2023.3
PyYAML==6.0.1
rest-framework-simplejwt==0.0.2
sqlparse==0.4.4
text-unidecode==1.3
tomli==2.0.1
typing_extensions==4.7.1
uritemplate==4.1.1
